Simple Jamaican Restaurant
Jamaican Restaurant that has Breakfast, Lunch and Dinner menus, each
menu has its own webpage, that can record multiply  orders. A form is
add on each page for the client who order to fill out information, each
line in the for has to valid before the form can be submitted.